chore: move html-rendering lib to @scalar/core
