+++ integrations/express/package.json
@@ -44,7 +44,8 @@
   ],
   "module": "dist/index.js",
   "dependencies": {
-    "@scalar/api-reference": "workspace:*"
+    "@scalar/core": "workspace:*",
+    "@scalar/types": "workspace:*"
   },
   "devDependencies": {
     "@scalar/build-tooling": "workspace:*",
