+++ integrations/express/src/apiReference.ts
@@ -1,6 +1,6 @@
+import { getHtmlDocument } from '@scalar/core/libs/html-rendering'
 import type { Request, Response } from 'express'
 
-import { getHtmlDocument } from '@scalar/api-reference/lib/html-rendering'
 import type { ApiReferenceConfiguration } from './types.ts'
 
 /**
