+++ integrations/express/src/types.ts
@@ -1,4 +1,4 @@
-import type { HtmlRenderingConfiguration } from '@scalar/api-reference/lib/html-rendering'
+import type { HtmlRenderingConfiguration } from '@scalar/core/libs/html-rendering'
 
 /**
  * The configuration for the Scalar API Reference for Express
