+++ integrations/fastify/package.json
@@ -49,26 +49,24 @@
   ],
   "module": "./dist/index.js",
   "dependencies": {
-    "@scalar/api-reference": "workspace:*",
-    "@scalar/types": "workspace:*",
-    "fastify-plugin": "^4.5.1"
+    "@scalar/core": "workspace:*",
+    "@scalar/openapi-types": "workspace:*",
+    "@scalar/openapi-parser": "workspace:*",
+    "fastify-plugin": "^4.5.1",
+    "github-slugger": "^2.0.0"
   },
   "devDependencies": {
     "@fastify/basic-auth": "^5.1.1",
     "@fastify/http-proxy": "^9.0.0",
     "@fastify/swagger": "^8.10.1",
     "@scalar/build-tooling": "workspace:*",
-    "@scalar/openapi-parser": "workspace:*",
     "@vitest/coverage-v8": "^1.6.0",
     "fastify": "^4.26.2",
-    "github-slugger": "^2.0.0",
-    "magic-string": "^0.30.8",
     "rollup-plugin-node-externals": "^7.1.2",
     "terser": "^5.30.0",
     "vite": "^5.4.10",
     "vite-plugin-static-copy": "^1.0.2",
     "vitest": "^1.6.0",
-    "vue": "^3.5.12",
     "yaml": "^2.4.5"
   }
 }
