+++ integrations/hono/package.json
@@ -42,7 +42,7 @@
   ],
   "module": "dist/index.js",
   "dependencies": {
-    "@scalar/api-reference": "workspace:*"
+    "@scalar/core": "workspace:*"
   },
   "devDependencies": {
     "@hono/node-server": "^1.11.0",
