+++ packages/api-reference/package.json
@@ -47,8 +47,7 @@
   "exports": {
     ".": "./dist/index.js",
     "./style.css": "./dist/style.css",
-    "./browser/standalone.js": "./dist/browser/standalone.js",
-    "./lib/html-rendering": "./dist/standalone/lib/html-rendering.js"
+    "./browser/standalone.js": "./dist/browser/standalone.js"
   },
   "files": [
     "dist",
