+++ packages/core/src/index.ts
@@ -1,0 +1,1 @@
+// Let’s keep this file empty, but use entrypoints for each lib.
