+++ packages/core/src/libs/html-rendering/index.ts
@@ -1,4 +1,4 @@
-import { apiReferenceConfigurationSchema, type ApiReferenceConfiguration } from '@scalar/types/api-reference'
+import { type ApiReferenceConfiguration, apiReferenceConfigurationSchema } from '@scalar/types/api-reference'
 import { z } from 'zod'
 
 /**
